requirements.txt
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.3
Markdown==3.1
ipdb==0.12
celery==4.3.0